==> studies/PCAWG/7_capabilities.qmd <==
---
title: "Demonstrating bayesNMF capabilities with PCAWG breast adenocarcinoma data example"
author: "Jenna Landy"
---

```{r}
library(bayesNMF)
library(ggplot2)
library(tidyverse)
reference <- get_cosmic()

output_dir = "../../output/PCAWG/bayesNMF"
figures_dir = "../../figures/capabilities/PCAWG"
```

```{r}
sampler <- readRDS(file.path(output_dir, "Breast_AdenoCA/sampler.rds"))
sampler$specs$output_dir = figures_dir
```

```{r}
trace_plot(sampler)
ggsave(file.path(figures_dir, "trace_plot.png"), height = 10, width = 12)

trace_plot(sampler, MAP_means = TRUE)
ggsave(file.path(figures_dir, "trace_plot_MAP.png"), height = 13, width = 12)
```

Note: Rownames of PCAWG data overlap but are not in the same order as the rownames of the COSMIC reference. However, bayesNMF auto-aligns rownames of reference to the rownames of the data. So we don't need to manually align the reference to the data.

```{r}
plot(sampler, reference, sigs = TRUE)
```

```{r}
sampler$reference_comparison$assignments
sampler$reference_comparison$votes
```
